rename cli
